% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{analyse_type_etablissement}
\alias{analyse_type_etablissement}
\title{Analyse etablissement}
\usage{
analyse_type_etablissement(es)
}
\arguments{
\item{es}{dataframe RPU (es = établissement de santé)}
}
\value{
"n.passages", "n.age.ren", "n.inf1an", "n.inf15ans", "n.75ans", "n.cp.rens",
"n.etrangers", "n.lun", "n.mar", "n.mer", "n.jeu", "n.ven", "n.sam", "n.dim", 
"n.nuit", "n.pds", "n.h.rens", "n.trans.rens", "n.fo",
"n.heli", "n.perso", "n.smur", "n.vsav", "n.ambu", "n.ccmu.rens", "n.ccmu1", 
"n.ccmu2", "n.ccmu3", "n.ccmu4",
"n.ccmu5", "n.ccmuP", "n.ccmuD", "n.ccmu45", "n.sorties.conf", "mean.passage", 
"median.passage", "n.passage4", "n.hosp.passage4", "n.dom.passage4", "n.dom", 
"n.hosp", "n.transfert", "n.deces", "n.mode.sortie",
"n.mutation2"
}
\description{
fournit une liste d'indicateur à partir des données d'un établissement
             ou d'un groupe d'établissements. Voir rapport 2014: Analyse par type d'étblissement
}
\examples{
# es non SAMU, siège de SMUR
          # es <- dx[dx$FINESS \%in\% c("Wis","Hag","Sav","Sel","Col"),]
          # analyse_type_etablissement(es)
}

