% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{completude}
\alias{completude}
\title{taux de completude global.}
\usage{
completude(dx, calcul = "percent", tri = FALSE)
}
\arguments{
\item{dx}{Un dataframe}

\item{calcul}{2 options "percent" (défaut) ou "somme". Somme = nb de réponses
non nulles. Percent = \% de réponses non nulles.}

\item{tri}{si tri = TRUE (defaut) les colonnes sont triées par ordre croissant.}
}
\value{
vecteur des taux de complétude
complétude brute. Des corrections sont nécessaires pour DESTINATION
correction pour Destination et Orientation
Les items DESTINATION et ORIENTATION ne s'appliquent qu'aux patients hspitalisés. 
On appelle hospitalisation les RPU pour lequels la rubrique MODE_SORTIE = MUTATION ou TRANSFERT. 
Pour les sorties à domicile, ces rubriques ne peuvent pas être complétées ce qui entraine 
une sous estimation importante du taux de complétude pour ces deux rubriques. 
On ne retient donc que le sous ensemble des patients hospitalisés pour lesquels les rubriques 
DESTINATION et ORIENTATION doivent être renseignées.
Correction pour DP. Cette rubrique ne peut pas être remplie dans le cas où ORIENTATION =
FUGUE, PSA, SCAM, REO
completude <- completude[-c(1,7)]
}
\description{
Pour chacune des rubriques RPU calcule le taux de réponse (complétude)
}
\details{
todo
}
\author{
JcB 2013-02-01
}
\seealso{
Other RPU: \code{\link{radar_completude}}
}
\keyword{complétude}

