% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{datetime}
\alias{datetime}
\title{met une string date au format YYYY-MM-DD HH:MM:SS}
\usage{
Transforme des rubriques ENTREE et SORTIE en objet datetime
}
\arguments{
\item{date}{une chaine de caractère de type Date}
}
\value{
un vecteur date time (lubridate)
}
\description{
met une string date au format YYYY-MM-DD HH:MM:SS
}
\note{
nécessite lubridate
}
\seealso{
horaire, passage.nuit
}

