% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{df.duree.pas.old}
\alias{df.duree.pas.old}
\title{Dataframe Durée de passage}
\usage{
df.duree.pas(dx, unit = "mins", mintime = 0, maxtime = 3)
}
\arguments{
\item{dx}{un dataframe de type RPU}

\item{unit}{unité de temps. Défaut = mins}

\item{mintime}{défaut = 0. Durée de passage minimale}

\item{maxtime}{défaut = 3 (72 heures). Durée de passage maximale}
}
\value{
dataframe de type duree_passage
}
\description{
fabrique à partir d'un dataframe de type RPU, un dataframe de type duree_passage comportant les colonnes suivantes:
date/heure d'entree, date/heure de sortie, durée de passage (en minutes par défaut), l'heure d'entrée (HMS), l'heure de sortie
}
\examples{
df <- df.duree.pas(dx)
}

