% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{duree.passage2}
\alias{duree.passage2}
\title{Calcul de la durée de passage}
\usage{
duree.passage2(dx, h1 = 0, h2 = 4320, hors_uhcd = TRUE)
}
\arguments{
\item{dx}{dataframe RPU}

\item{h1}{durée minimale en minutes (par défaut > 0)}

\item{h2}{durée maximale en minutes (par défaut 4320 = 72 heures)}

\item{hors_uhcd}{si TRUE (défaut) on retire les engegistrements où ORIENTATION = UHCD}
}
\value{
dataframe à 4 colonnes: entree, sortie, mode_sortie, duree (en mn),
                                he (heure d'entrée), hs (heure de sortie)
}
\description{
todo
}

