% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{evolution}
\alias{evolution}
\title{Evolution d'une années sur l'autre}
\usage{
evolution(n.rpu, n.rpu.2013)
}
\arguments{
\item{a}{chiffre de l'année courante}

\item{b}{chiffre de l'année précédente}
}
\value{
pourcentage d'augmentation ou de diminution
}
\description{
calcule l'évolution entre 2 chiffres
}

