% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{finess2territoires}
\alias{finess2territoires}
\title{réorganiser les FINESS par territoires de santé}
\usage{
finess2territoires(finess)
}
\arguments{
\item{finess}{code finess de létablissement}
}
\examples{
dx$FINESS <- finess2territoires(dx)
}

