% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{format.n}
\alias{format.n}
\title{formate un nombre}
\usage{
format.n(7890.14) -> "7 890,14"
}
\description{
formate un nombre en ajoutant un espace pour les milliers
                                          une virgule décimale
                                          pas de notation scientifique
                                          deux chiffres significatifs
}
\author{
JcB - 2015-03-12
}

