% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{imprime.resume.rpu}
\alias{imprime.resume.rpu}
\title{Imprime un summary.rpu}
\usage{
imprime.resume.rpu(x, sens = "colonne", cnames = NULL, rnames = NULL,
                         caption = "", type = "latex", ref = "")
}
\arguments{
\item{x}{un vecteur nommé}

\item{sens}{'colonne' = vertical, 'ligne' = horizontal}

\item{cnames}{noms des colonnes}

\item{rnames}{noms des lignes}
}
\description{
imprime un objet de type summary.rpu, en ligne eou en colonne (défaut) avec xtable.
}
\examples{
x <- ummary.wday(es$ENTREE))
       print.summary.rpu(x, cnames = c("Jour","n"), caption = "Nombre de RPU par jour de semaine")
}

