% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{is.cpals}
\alias{is.cpals}
\title{Est-ce un code postal alsacien ?}
\usage{
is.cpals(cp)
}
\arguments{
\item{cp}{code postal ou un vecteur de codes postaux}
}
\value{
TRUE ou FALSE
}
\description{
Détermine si un CP appartient à la région Alsace
}
\examples{
 is.cpals(as.factor("67550")), is.cpals("56000")
a <- is.cpals(dx$CODE_POSTAL)

}

