% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{isWE}
\alias{isWE}
\title{Determine si la date/heure correspond à un week-end.}
\usage{
isWE(date)
}
\arguments{
\item{date}{date/heure de type YYYY-MM-DD HH:MM:SS}
}
\value{
boolean
}
\description{
retourne TRUE si on est en horaire de week-end et False sinon
}
\details{
la période de WE s'étend du vendredi 20 heures au lundi 8 heures. Nécessite lubridate. 
Ne traite qu'une date à la fois.
}
\examples{
isWE("2015-12-28 05:12:00") # TRUE
          isWE(as.Date("2015-12-28 05:12:00")) # FALSE
}

