% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{mn2h}
\alias{mn2h}
\title{transforme des minutes en heure/mn}
\usage{
mn2h(x)
}
\arguments{
\item{x}{integer = nombre de minutes}
}
\value{
char
}
\description{
transforme des minutes en heure/mn
}

