% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{passages2}
\alias{passages2}
\title{Nombre de RPU sur une plage horaire donnée}
\usage{
n.passages.nuit <- passages2(pop18$ENTREE, "nuit")
}
\arguments{
\item{vx}{vecteur de type datetime (dx$ENTREE, dx$SORTIE par exemple). Transformé par ymd_hms
Transform dates stored as character or numeric vectors to POSIXct objects}

\item{h1}{char heure de début ou période: 'nuit', nuit_profonde', 'jour', 'pds',
'soir', '08:00:00'}

\item{h2}{char heure de fin. h2 doit être > h1}
}
\value{
integer
}
\description{
Détermine le nombre de RPU sur une plage horaire donnée
}
\author{
jcb
}

