% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{plot.xts2}
\alias{plot.xts2}
\title{plot.xts en couleur}
\source{
http://stackoverflow.com/questions/9017070/set-the-color-in-plot-xts
}
\usage{
plot.xts2(x, y = NULL, type = "l", auto.grid = TRUE, major.ticks = "auto", minor.ticks = TRUE, major.format = TRUE, 
bar.col = "grey", candle.col = "white", ann = TRUE, axes = TRUE, col = "black", ...)
}
\description{
La méthode plot.xts comprte un bug qui empêche l'affichage de courbes en couleur. Cette version corrige le bug.
}
\author{
Roman Luštrik (http://stackoverflow.com/users/322912/roman-lu%c5%a1trik)
}

