% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{print.summary.rpu}
\alias{print.summary.rpu}
\title{Imprime un summary.rpu}
\usage{
print.summary.rpu(x, sens = "colonne", cnames = NULL, rnames = NULL,
                         caption = "", type = "latex", ref = "")
}
\arguments{
\item{x}{un vecteur nommé}

\item{sens}{'colonne' = vertical, 'ligne' = horizontal}

\item{cnames}{noms des colonnes}

\item{rnames}{noms des lignes}
}
\description{
imprime un objet de type summary.rpu, en ligne eou en colonne (défaut).
}
\details{

}
\examples{
x <- ummary.wday(es$ENTREE))
       print.summary.rpu(x, names = c("Jour","n"), caption = "Nombre de RPU par jour de semaine")
}

