% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{print_table_rpu}
\alias{print_table_rpu}
\title{Imprime une table avec xtable.}
\usage{
print_table_rpu(t, caption = "", type = "latex", ref = "")
}
\arguments{
\item{t}{un objet de type table}

\item{caption}{une légende. Mettre c("légende", "sommaire") si nécessaire}

\item{type}{"latex" ou "html"}

\item{ref}{référence du tableau (latex)}
}
\description{
imprime une table avec xtable. Par défaut l'environnement est du type latex, le
             séparateur de milliers est l'espace et la virgule décimale
}
\examples{
  print_table_rpu(t)
         print_table_rpu(t, "table de test")
         print_table_rpu(t, "table de test", "html")
}

