% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{radar.completude}
\alias{radar.completude}
\title{dessine un graphe en étoile}
\usage{
radar.completude(completude(dx))
}
\arguments{
\item{completude}{taux de completude global calculé par la fonction completude}

\item{finess}{character: nom de l'établissement. NULL (defaut) => tout le datafame}
}
\value{
diagramme en étoile
}
\description{
dessine un graphe en étoile à partir des données retournées par "completude"
}
\author{
JcB 2013-02-01
}
\seealso{
Other RPU: \code{\link{completude}}
}
\keyword{diagramme}
\keyword{spider,}
\keyword{étoile}

