% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{reorder.dataframe.fedoru}
\alias{reorder.dataframe.fedoru}
\title{Reordonne les colonnes du dataframe RPU dans l'ordre defini par la FEDORU.}
\usage{
reorder.dataframe.fedoru(dx)
}
\arguments{
\item{dx}{un dataframe de type RPU}
}
\description{
Permet une meilleure cohérence du diagramme en étoile
}

