% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{resume.age.sexe}
\alias{resume.age.sexe}
\title{résumé des vecteurs AGE et SEXE}
\usage{
summary.age.sexe(dx)
}
\arguments{
\item{dx}{dataframe RPU}
}
\value{
moyenne, écart-type, médiane par sexe
}
\description{
résumé des vecteurs AGE et SEXE
}
\examples{
summary.age.sexe(dx)
}

