% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{resume.dateheure}
\alias{resume.dateheure}
\title{Resume du vecteur des ENTREE ou SORTIE}
\usage{
summary.dateheure(vx)
}
\arguments{
\item{vx}{vecteur ENTREE ou SORTIE}
}
\value{
"n", "n.na", "p.na", "n.rens", "p.rens"
}
\description{
résumé du vecteur vx des ENTREE ou SORTIE
}
\examples{
summary.ccmu(dx$SORTIE)
}

