% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{resume.passages}
\alias{resume.passages}
\title{analyse un objet de type duree.passage2}
\usage{
summary.passages(dp)
}
\arguments{
\item{dp}{un objet de type duree.passage2. Correspond à un dataframe d'éléments du RPU dont
la rurée de passage est conforme cad non nulle et inférieure à 72 heures}
}
\value{
n.conforme               NB de durées conformes (>0 mn et < 72 heures)
        duree.moyenne.passage    durée moyenne d'un passage en minutes
        duree.mediane.passage    durée médiane d'un passage en minutes
        duree.moyenne.passage.dom    durée moyenne d'un passage en minutes si retour dom
        duree.mediane.passage.dom    durée médiane d'un passage en minutes
        duree.moyenne.passage.hosp    durée moyenne d'un passage en minutes si hospit.
        duree.mediane.passage.hosp    durée médiane d'un passage en minutes
        n.passage4               nombre de passages de moins de 4 heures
        n.hosp.passage4          nombre de passages de moins de 4 heures suivi d'hospitalisation
        n.domicile               nombre de retours à domicile
        n.dom.passage4           nombre de passages de moins de 4 heures suivi d'un retour à domicile
        n.dom                    nombre de retours à domicile
}
\description{
analyse un objet de type duree.passage2
}

