% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{resume.transport}
\alias{resume.transport}
\title{analyse du vecteur TRANSPORT}
\usage{
summary.transport(vx)
}
\arguments{
\item{vx}{vecteur de Factor}
}
\value{
"n", "n.na", "p.na", "n.rens", "p.rens", "n.fo", "n.heli", "n.perso", "n.smur",
         "n.vsav", "n.ambu", "p.fo", "p.heli", "p.perso", "p.smur", "p.vsav", "p.ambu"
}
\description{
analyse du vecteur TRANSPORT
}
\examples{
summary.transport(pop75$TRANSPORT)
}

