% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{rpu.par.jour}
\alias{rpu.par.jour}
\title{Nombre de RPU par jour et par FINESS}
\usage{
rpu.par.jour(dx)

rpu.par.jour(d, roll = 7)
}
\arguments{
\item{d}{vecteur de dates compatible avec le format Date}

\item{dx}{un dataframe de type rpu ayant un minimum 2 colonnes ENTREE et FINESS}

\item{roll:}{nb de jours pour la moyenne lissée. Défaut = 7}
}
\value{
un dataframe de 4 colonnes: date calendaire, nb de RPU du jour, le n° du jour de l'année (1 à 365), la moyennne lissée
}
\description{
retourne une table contenant le nombre de RPU par jour et par FINESS

A partir d'un vecteur de dates, calcule le nombre de RPU par jour
}
\details{
RAJOUTER LES SOMMES   CUMuLEES. Nécessite xts, lubridate
}
\examples{
rpu.par.jour(d04)
p2013 <- rpu.par.jour(j2013$ENTREE)
       plot(p2013$V2, type="l") # les RPU
       lines(p2013$V3, p2013$V4) # moyenne mobile
}

