% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{rpu.par.mois}
\alias{rpu.par.mois}
\title{Nombre de RPU par mois}
\usage{
rpu.par.mois(dx, standard = FALSE)
}
\arguments{
\item{dx}{dataframe (au minimum la colonne ENTREE)}

\item{standard}{(boolean) si true retourne par mois corrigés de 30j sinon le nombre brut de RPU}
}
\value{
un vecteur nommé: nom du mois, nb de RPU
}
\description{
Calcule le nombre de RPU par mois entre deux dates sous forme brute
ou corrigée en mois constants de 30 jours.
}
\examples{
tc1 <- rpu.par.mois(d15, FALSE)
tc2 <- rpu.par.mois(d15, TRUE)
a <- rbind(tc1, tc2)
par(mar=c(5.1, 4.1, 8.1, 2), xpd=TRUE)
barplot(a, beside = TRUE, cex.names = 0.8)
legend("topleft", inset = c(0, -0.1), legend = c("Brut","Standardisé"), bty = "n", col = c("black","gray80"), pch = 15)
}

