\name{Rpu2-package}
\alias{Rpu2-package}
\alias{Rpu2}
\docType{package}
\title{
\packageTitle{Rpu2}
}
\description{
\packageDescription{Rpu2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rpu2}
\packageIndices{Rpu2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Rpu2}

Maintainer: \packageMaintainer{Rpu2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
