% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{summary.age.sexe}
\alias{summary.age.sexe}
\title{résumé des vecteurs AGE et SEXE}
\usage{
summary.age.sexe(dx)
}
\arguments{
\item{dx}{dataframe RPU}
}
\value{
moyenne, écart-type, médiane par sexe
}
\description{
résumé des vecteurs AGE et SEXE
}

