% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{summary.ccmu}
\alias{summary.ccmu}
\title{résumé du vecteur vx des CCMU}
\usage{
summary.ccmu(dx$GRAVITE)
}
\arguments{
\item{vx}{vecteur de factor CCMU}
}
\value{
"n", "n.na", "p.na", "n.rens", "p.rens", "n.ccmu1", "n.ccmu2", "n.ccmu3",
"n.ccmu4", "n.ccmu5", "n.ccmup", "n.ccmud", "p.ccmu1", "p.ccmu2", "p.ccmu3", "p.ccmu4", "p.ccmu5", "p.ccmup", "p.ccmud")
}
\description{
résumé du vecteur vx des CCMU
}

