% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{summary.cp}
\alias{summary.cp}
\title{résumé du vecteur CODE_POSTAL (cp)}
\usage{
summary.cp(dx$CODE_POSTAL)
}
\arguments{
\item{vx}{vecteur char CODE_POSTAL}
}
\value{
- nb de CP renseignés
         - nb de résidents alsaciens
         - nb d'étrangers
}
\description{
résumé du vecteur vx des CODE_POSTAL (cp)
}
\details{
NECESSITE LA BIBLIOTHEQUE RPU_Doc/mes.constantes
}

