% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{summary.dateheure}
\alias{summary.dateheure}
\title{résumé du vecteur des ENTREE ou SORTIE}
\usage{
summary.ccmu(dx$SORTIE)
}
\arguments{
\item{vx}{vecteur ENTREE ou SORTIE}
}
\value{
"n", "n.na", "p.na", "n.rens", "p.rens"
}
\description{
résumé du vecteur vx des ENTREE ou SORTIE
}

