% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{summary.destination}
\alias{summary.destination}
\title{Resume de la DESTINATION}
\usage{
summary.destination(dx, correction = TRUE)
}
\arguments{
\item{dx}{dataframe RPU}

\item{correction}{= TRUE: on ne retient que les destinations 
correspondant à une hospitalisation}
}
\value{
"n", "n.na", "p.na", "n.rens", "p.rens"
}
\description{
résumé du vecteur vx des DESTINATION. En cas d'hospitalisation, il y a quatre destinations possibles:
MCO, SSR, SLD et PSY. En ca  de sortie au domicile: HAD et Structure médico-sociale (EHPAD)
}
\details{
MANQUE LE SUMMARY DU VECTEUR.
}

