% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rapport.R
\name{summary.duree.passage}
\alias{resume.duree.passage}
\alias{summary.duree.passage}
\title{Resume de la Duree de passage.}
\usage{
summary.duree.passage(dp)
}
\arguments{
\item{dp}{un objet de type duree.passage2}
}
\value{
- nb de durées
        - min durée
        - max durée
        - durée moyenne
        - durée médiane
        - écart-type
        - 1er quartile
        - 3ème quartile
}
\description{
Résumé de dp. dp est produit par duree.passages2 et se présente sous forme d'un 
data.frame à 4 colonnes

analyse de la colonne durée
}

