% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{summary.mode.sortie}
\alias{summary.mode.sortie}
\title{résumé du vecteur vx des MODE_SORTIE}
\usage{
summary.mode.sortie(dx$MODE_SORTIE)
}
\arguments{
\item{vx}{vecteur char MODE_SORTIE}
}
\value{
"n", "n.na", "p.na", "n.rens", "p.rens",
"n.dom", "n.hosp", "n.transfert", "n.mutation", "n.deces", "p.dom", "p.hosp", "p.transfert", "p.mutation", "p.deces")
}
\description{
résumé du vecteur vx des MODE_SORTIE
}

