% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{summary.orientation}
\alias{summary.orientation}
\title{Resume de ORIENTATION}
\usage{
summary.orientation(dx, correction = TRUE)
}
\arguments{
\item{dx}{dataframe RPU}

\item{correction}{= TRUE: on ne retient que les orientation
correspondant à une hospitalisation}
}
\value{
"n", "n.na", "p.na", "n.rens", "p.rens",
"n.chir", "n.med", "n.obst", "n.si", "n.sc", "n.rea", "n.uhcd", "n.ho", "n.hdt",
"n.reo", "n.scam", "n.psa",
"p.chir", "p.med", "p.obst", "p.si", "p.sc", "p.rea", "p.uhcd", "p.ho", "p.hdt",
"p.reo", "p.scam", "p.psa"
}
\description{
résumé du vecteur vx des ORIENTATION
}

