% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{summary.rpu}
\alias{summary.rpu}
\title{calcule le nombre de RPU par SU, territoire de sante et
             departement.}
\source{
summary_rpu.R
}
\usage{
summary.rpu(dx)
}
\arguments{
\item{dx}{un dataframe RPU ou un dataframe réduit à 2 colonnes: ENTREE et
FINESS}
}
\value{
un objet "list"
       n nombre total de RPU
       n.tx  total RPU du territoire x
       n.67  total pour le 67
       n.68  total pour 68
       n.xxx total pour le Finess xxx
       p.tx  % pour territoire x
}
\description{
calcule le nombre de RPU par SU, territoire de santé et
             département à partir d'un dataframe RPU. Deux colonnes sont
             indispensables: ENTREE et FINESS
}
\details{
v1.0 24/08/2015
}
\examples{
s <- summary.rpu(d15); s[1]; s$debut; s$n
}
\author{
JcB - 2015-08-24
}

