% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{synthese.completude}
\alias{synthese.completude}
\title{Calcule le tableau des taux de completude de l'ensemble des Finess.}
\usage{
synthese.completude(dx)
}
\arguments{
\item{dx}{dataframe de type RPU}
}
\value{
un dataframe
}
\description{
A partir du dataframe initial (dx) calcule le tableau des taux de complétude
de l'ensemble des Finess présents dans dx.
}
\details{
Le tableau comporte en ordonnée le
nom des établissements, en abcisse les différents items du RPU et à l'intersection
ligne/colonne la complétude correspondante. dx peut comprter un ou plusieurs Finess
et concerner une période variable (semaine, mois, année...)
Nécessite la librairie plyr pour la fonction ddply()
}
\examples{
synthese.completude(dx)
synthese.completude(dx[dx$FINESS == "Hag",]) pour un seul établissement
}

