% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{week.variations}
\alias{week.variations}
\title{Variation du nombre de RPU par semaine}
\usage{
week.variations(vx, last = FALSE)
}
\arguments{
\item{vx}{vecteur du nombre de RPU pr semaine (voir week.rpu)}

\item{last}{boolean Si TRUE, on élimine la dernière semaine qui est souvent incomplète. FALSE par défaut.}
}
\value{
un vecteur d'entiers positifs ou négatifs
}
\description{
Variation du nombre de RPU par semaine
}
\examples{
# d3 <- week.rpu(dx[dx$FINESS == "3Fr",])
# v <- week.variations(d3)
}

