% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{passages2}
\alias{passages2}
\title{Nombre de RPU sur une plage horaire donnee}
\usage{
passages2(vx, h1, h2 = NULL)
}
\arguments{
\item{vx}{vecteur de type datetime (dx$ENTREE, dx$SORTIE par exemple). Transformé par ymd_hms
Transform dates stored as character or numeric vectors to POSIXct objects}

\item{h1}{char heure de début ou période: 'nuit', nuit_profonde', 'jour', 'pds',
'soir', '08:00:00'}

\item{h2}{char heure de fin. h2 doit être > h1}
}
\value{
2 objets: nombre de RPU et pourcentage
}
\description{
Détermine le nombre de RPU sur une plage horaire donnée et le pourcentage par rapport
au nombre total de passages contenus dans vx.
}
\details{
nécessite lubridate library(lubridate)
}
\examples{
n.passages.nuit <- passages2(pop18$ENTREE, "nuit"); n.passages.nuit[1]; n.passages.nuit[2]
}
\author{
jcb
}

