% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{reorder.dataframe.fedoru}
\alias{reorder.dataframe.fedoru}
\title{Réordonne les colonnes du dataframe RPU dans l'ordre défini par la FEDORU.}
\usage{
\method{reorder}{dataframe.fedoru}(dx)
}
\arguments{
\item{dx}{un dataframe de type RPU}
}
\description{
Permet une meilleure cohérence du diagramme en étoile
}

