% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{summary.dp}
\alias{summary.dp}
\title{Resume du vecteur DP (diagnostic principal)}
\usage{
summary.dp(vx)
}
\arguments{
\item{vx}{vecteur char DP}
}
\value{
"n", "n.na", "p.na", "n.rens", "p.rens"
}
\description{
résumé du vecteur vx des DP (diagnostic principal)
}
\examples{
summary.dp(dx$DP)
}

