% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rapport.R
\name{teste.radar}
\alias{teste.radar}
\title{data pour créer automatiquement un radar RPU et faire des test}
\usage{
teste.radar()
}
\description{
data pour créer automatiquement un radar RPU et faire des test
}
\examples{
teste.radar()
}

